presto/dataops/requirements.txt
google-cloud-storage
numpy
openmapflow[data]==0.2.1
